Use apm ci

+++ build-package.ps1
@@ -66,20 +66,45 @@ function PrintVersions() {
 
 function InstallPackage() {
     Write-Host "Downloading package dependencies..."
-    & "$script:APM_SCRIPT_PATH" clean
-    if ($LASTEXITCODE -ne 0) {
-        ExitWithCode -exitcode $LASTEXITCODE
-    }
     if ($script:ATOM_LINT_WITH_BUNDLED_NODE -eq $TRUE) {
-      & "$script:APM_SCRIPT_PATH" install
+      if (Test-Path -Path "package-lock.json" -PathType Leaf) {
+        & "$script:APM_SCRIPT_PATH" ci
+        if ($LASTEXITCODE -ne 0) {
+          ExitWithCode -exitcode $LASTEXITCODE
+        }
+      } else {
+        Write-Warning "package-lock.json not found; running apm install instead of apm ci"
+        & "$script:APM_SCRIPT_PATH" install
+        if ($LASTEXITCODE -ne 0) {
+          ExitWithCode -exitcode $LASTEXITCODE
+        }
+
+        & "$script:APM_SCRIPT_PATH" clean
+        if ($LASTEXITCODE -ne 0) {
+          ExitWithCode -exitcode $LASTEXITCODE
+        }
+      }
       # Set the PATH to include the node.exe bundled with APM
       $newPath = "$script:PACKAGE_FOLDER\$script:ATOM_DIRECTORY_NAME\resources\app\apm\bin;$env:PATH"
       $env:PATH = $newPath
       [Environment]::SetEnvironmentVariable("PATH", "$newPath", "User")
     } else {
-      & "$script:APM_SCRIPT_PATH" install --production
-      if ($LASTEXITCODE -ne 0) {
+      if (Test-Path -Path "package-lock.json" -PathType Leaf) {
+        & "$script:APM_SCRIPT_PATH" ci --production
+        if ($LASTEXITCODE -ne 0) {
+            ExitWithCode -exitcode $LASTEXITCODE
+        }
+      } else {
+        Write-Warning "package-lock.json not found; running apm install instead of apm ci"
+        & "$script:APM_SCRIPT_PATH" install --production
+        if ($LASTEXITCODE -ne 0) {
+            ExitWithCode -exitcode $LASTEXITCODE
+        }
+
+        & "$script:APM_SCRIPT_PATH" clean
+        if ($LASTEXITCODE -ne 0) {
           ExitWithCode -exitcode $LASTEXITCODE
+        }
       }
       # Use the system NPM to install the devDependencies
       Write-Host "Using Node.js version:"
